+++ src/app/(auth)/login/types/auth.ts
@@ -1,0 +1,5 @@
+export enum PATHS {
+  HOME = "/",
+  LOGIN = "/login",
+  REGISTER = "/register",
+}
