--- src/types/auth.ts
@@ -1,5 +1,0 @@
-export enum PATHS {
-  HOME = "/",
-  LOGIN = "/login",
-  REGISTER = "/register",
-}
