--- src/types/models/notification.types.ts
@@ -1,8 +1,0 @@
-export interface Notification {
-  id: number
-  type: 'post' | 'comment' | 'like' | 'mention'
-  message: string
-  isRead: boolean
-  createdAt: string
-  link?: string
-}
