--- src/types/user.d.ts
@@ -1,9 +1,0 @@
-export interface User {
-  id: number
-  name: string
-  email: string
-  role: 'ADMIN' | 'USER'
-  createdAt: string
-  updatedAt: string
-}
-
